Added string based actor
